Nova 12.1 RC1.1 Build

+++ $OEM$/$$/Setup/Scripts/Run/Clean.ps1
@@ -1,9 +1,20 @@
+$ScriptDir = Split-Path $script:MyInvocation.MyCommand.Path
 # Clean up Script
 
-# Remove startup items, set UAC back to normal, and Remove InstallVar
+# Load Variables
+if (Test-path "$env:windir\Setup\Scripts"){cd $env:windir\Setup\Scripts\Run} else {cd $ScriptDir}
+. .\GlobalVars.ps1
+
+# Copy Setup Log to User Folder
+if (Test-path Setup_Update.log){mv "Setup_Update.log" "$env:UserProfile"}
+
+# Set UAC back to normal
 New-ItemProperty "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\System" -Name EnableLUA -Value 1 -PropertyType "DWORD" -Force | out-null
-Rename-Item "$env:programdata\Microsoft\Windows\Start Menu\Programs\StartUp\Starter.bat" "temp.TMP"
-Remove-Item "$env:programdata\Microsoft\Windows\Start Menu\Programs\StartUp\temp.TMP"
+
+# Remove startup items if necessary
+if (Test-path "$Startup\Starter.bat"){ 
+	Rename-Item "$Startup\Starter.bat" "temp.TMP"
+	Remove-Item "$Startup\temp.TMP"}
 
 # Fix Run Once for Vista / Server 2008
 if ($winver -like "6.0.*"){
@@ -15,10 +26,10 @@ cd "$default\Server\2008r2"
 start-process "Packs\W7packsR2.exe" -ArgumentList "/silent" -wait}
 
 # Delete Scripts Folder & PowerShell Profile
-cd $env:temp
-sc wipe.ps1 'rm -r -force "$env:windir\Setup\Scripts"'
-ac wipe.ps1 'rm -r -force $env:userprofile\documents\WindowsPowerShell'
-. .\wipe.ps1
-
+if (test-path "$env:windir\Setup\Scripts"){
+	cd $env:temp
+	sc wipe.ps1 'rm -r -force "$env:windir\Setup\Scripts"'
+	. .\wipe.ps1}
+	
 Restart-Computer -Force
 exit
